+++ CHANGES.rst
@@ -4,7 +4,10 @@ Changelog of lizard-nxt client
 Unreleased
 ----------
 
--
+- Events series in timeline are colored.
+
+- Event series can be differentiated by color.
+
 
 Release 0.2b10 (2014-11-06)
 ---------------------------